+++ app/src/main/java/ru/gb/veber/toplibrary/core/ConnetivityListener.kt
@@ -3,11 +3,9 @@ package com.mirkhusainov.geekbrainscourse.core.utils
 import android.net.ConnectivityManager
 import android.net.Network
 import android.net.NetworkRequest
-import android.util.Log
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.core.Single
 import io.reactivex.rxjava3.subjects.BehaviorSubject
-import io.reactivex.rxjava3.subjects.PublishSubject
 
 class ConnetivityListener(connectivityManager: ConnectivityManager) {
