+++ app/src/main/java/ru/gb/veber/toplibrary/utils/RxJava.kt
@@ -7,9 +7,6 @@ import io.reactivex.rxjava3.core.Single
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 import io.reactivex.rxjava3.disposables.Disposable
 import io.reactivex.rxjava3.schedulers.Schedulers
-import ru.gb.veber.toplibrary.model.data.GithubUser
-import ru.gb.veber.toplibrary.model.data.ReposDto
-
 
 
 fun <T> Single<T>.doCompletableIf(
