+++ app/src/main/java/ru/gb/veber/toplibrary/view/userdetails/ReposAdapter.kt
@@ -44,6 +44,7 @@ class GithubUserReposViewHolder(
 
         dateCreating.text=item.createdAt
 
+        languageRepo.text=item.language
         root.setOnClickListener {
            onUserClickListener.invoke(item)
         }
