+++ app/src/main/res/layout/fragment_user_screen.xml
@@ -2,15 +2,16 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools">
 
     <ImageView
         android:id="@+id/ivUserAvatar"
         android:layout_width="100dp"
         android:layout_marginTop="28dp"
+        android:layout_marginLeft="28dp"
         android:layout_height="100dp"
         android:src="@drawable/ic_user_placeholder"
-        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -19,10 +20,23 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textSize="30sp"
-        android:layout_marginTop="42dp"
-        app:layout_constraintEnd_toEndOf="parent"
+        tools:text="UserName"
+        android:layout_marginLeft="28dp"
+        app:layout_constraintTop_toTopOf="@+id/ivUserAvatar"
+        app:layout_constraintBottom_toBottomOf="@+id/ivUserAvatar"
+        app:layout_constraintStart_toEndOf="@+id/ivUserAvatar"
+        />
+
+    <TextView
+        android:id="@+id/userRepos"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textSize="30sp"
+        tools:text="Repos:30"
+        android:padding="10dp"
+        app:layout_constraintTop_toBottomOf="@+id/ivUserAvatar"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/ivUserAvatar" />
+        />
 
     <ProgressBar
         android:id="@+id/progressBar"
@@ -39,9 +53,10 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="42dp"
+        android:padding="10dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/userName" />
+        app:layout_constraintTop_toBottomOf="@+id/userRepos" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
